+++ packages/insomnia/types/jsonlint.d.ts
